// src/interfaces/category.interface.ts
export enum Category {
  NO_RECONOCIDO = "NO RECONOCIDO",
  ENTERO = "ENTERO",
  REAL = "REAL",
  CARACTER = "CARACTER",
  LETRA = "LETRA",
  IDENTIFICADOR = "IDENTIFICADOR",
  PALABRA_RESERVADA = "PALABRA RESERVADA",
  CADENA_CARACTERES = "CADENA CARACTERES",
  COMENTARIO_LINEA = "COMENTARIO LINEA",
  COMENTARIO_BLOQUE = "COMENTARIO BLOQUE",
  OPERADOR_ARITMETICO = "OPERADOR ARITMETICO",
  OPERADOR_RELACIONAL = "OPERADOR RELACIONAL",
  OPERADOR_LOGICO = "OPERADOR LOGICO",
  OPERADOR_INCREMENTO_DECREMENTO = "OPERADOR INCREMENTO O DECREMENTO",
  OPERADOR_ASIGNACION = "OPERADOR ASIGNACION",
  COMENTARIO = "COMENTARIO",
  ERROR_SINTACTICO = "ERROR SINTACTICO",
  SEPARADOR = "SEPARADOR",
  PARENTESIS_CIERRE = "PARENTESIS DE CIERRE",
  PARENTESIS_APERTURA = "PARENTESIS DE APERTURA",
  LLAVES_APERTURA = "LLAVES APERTURA",
  LLAVES_CIERRE = "LLAVES CIERRE",
  HEXADECIMAL = "HEXADECIMAL",
  FIN_DE_SENTENCIA = "FIN DE SENTENCIA"
  }
  